# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integer Spiral API Documentation",
        "description": "Integer Spiral Project API's Request and Response Values",
        "contact": {
            "email": "hakan.syturk@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/createlayout": {
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Creates a layout with spiral form, takes 2 parameter and returns id of layout.",
                "description": "Creates a layout with spiral form, takes 2 parameter which are x and y. Returns id of created layout.",
                "operationId": "creatingLayout",
                "parameters": [
                    {
                        "name": "x",
                        "in": "query",
                        "description": "Size of x axis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "y",
                        "in": "query",
                        "description": "Size of y axis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getlayouts": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "List all layouts",
                "description": "Returns all informations of layouts.",
                "operationId": "gettingLayout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getvalue": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Returns coordinate according to x and y values",
                "description": "Returns the cell value of the layout according to the entered id, x and y coordinate values.",
                "operationId": "gettingValue",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of layout",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "x",
                        "in": "query",
                        "description": "x coordinate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "y",
                        "in": "query",
                        "description": "y coordinate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}